+++ haskell/glue/test/Glue/Lib/Math/Power/SqrtSpec.hs
@@ -16,21 +16,21 @@ spec = describe "Glue.Lib.Math.Power.Sqrt (Test sqrt function)" do
             result <- runEvalLegacy (Sqrt.sqrt args) (E.fromFrame lib)
             case result of
                 Left err -> expectationFailure $ "Sqrt failed: " <> show err
-                Right (res, _, _) -> res `shouldBe` Integer 2
+                Right (res, _, _) -> res `shouldBe` Float 2
 
         it "returns 3 for sqrt(9)" do
             let args = [Integer 9]
             result <- runEvalLegacy (Sqrt.sqrt args) (E.fromFrame lib)
             case result of
                 Left err -> expectationFailure $ "Sqrt failed: " <> show err
-                Right (res, _, _) -> res `shouldBe` Integer 3
+                Right (res, _, _) -> res `shouldBe` Float 3
 
         it "returns 0 for sqrt(0)" do
             let args = [Integer 0]
             result <- runEvalLegacy (Sqrt.sqrt args) (E.fromFrame lib)
             case result of
                 Left err -> expectationFailure $ "Sqrt failed: " <> show err
-                Right (res, _, _) -> res `shouldBe` Integer 0
+                Right (res, _, _) -> res `shouldBe` Float 0
 
         it "fails with negative numbers" do
             let args = [Float (-4)]
