+++ haskell/glue/test/Glue/Lib/Math/Trigonometric/CosSpec.hs
@@ -1,7 +1,6 @@
 module Glue.Lib.Math.Trigonometric.CosSpec (spec) where
 
 import Data.Either (isLeft)
-import Data.Scientific (fromFloatDigits, toRealFloat)
 import Glue.Env qualified as E
 import Glue.Eval (runEvalLegacy)
 import Glue.IR (IR (..))
@@ -17,7 +16,7 @@ spec = describe "Glue.Lib.Math.Trigonometric.Cos (Test cos function)" do
             result <- runEvalLegacy (Cos.cos args) (E.fromFrame lib)
             case result of
                 Left err -> expectationFailure $ "Cos failed: " <> show err
-                Right (res, _, _) -> res `shouldBe` Integer 1
+                Right (res, _, _) -> res `shouldBe` Float 1
 
         it "returns 0 for cos(π/2)" do
             let args = [Float (pi / 2)]
